Do not dispatch a searchAction if searchString === "" ?
Hi, 

First, thanks a lot for this amazing library. 

I'm using it in my app, and i'm running on this issue, if i am on an unrelated part of the app, where there is no search, and i update the entities of the index, the index get rebuilt. This is a normal behaviour. 

But in the redux dev tools i see a search action for each time the index get rebuilt, and this shouldn't happen as i'm not performing any search there.

What do you think ?